phase3_generate/requirements.txt
langchain-groq
python-dotenv
langchain-core
